move the OSDUtilizationDumper code into OSDMap
